refactor ReferencedByList: update import path for getReferencedBy utility function
