+++ src/utils/graph_utils.ts
